+++ openssl-sys/src/pem.rs
@@ -137,6 +137,14 @@ extern "C" {
         cb: pem_password_cb,
         u: *mut c_void,
     ) -> *mut EVP_PKEY;
+    pub fn d2i_PKCS8_PRIV_KEY_INFO(
+        k: *mut *mut PKCS8_PRIV_KEY_INFO,
+        buf: *mut *const u8,
+        length: c_long,
+    ) -> *mut PKCS8_PRIV_KEY_INFO;
+    pub fn PKCS8_PRIV_KEY_INFO_free(
+        p8inf: *mut PKCS8_PRIV_KEY_INFO,
+    );
 
     pub fn PEM_read_bio_PKCS7(
         bio: *mut BIO,
@@ -146,6 +154,16 @@ extern "C" {
     ) -> *mut PKCS7;
 
     pub fn PEM_write_bio_PKCS7(bp: *mut BIO, x: *mut PKCS7) -> c_int;
+
+    #[cfg(ossl101)]
+    pub fn PEM_read_bio_CMS(
+        bio: *mut BIO,
+        out: *mut *mut CMS_ContentInfo,
+        callback: pem_password_cb,
+        user_data: *mut c_void,
+    ) -> *mut CMS_ContentInfo;
+    #[cfg(ossl101)]
+    pub fn PEM_write_bio_CMS(bio: *mut BIO, cms: *const CMS_ContentInfo) -> c_int;
 }
 
 pub const PEM_R_NO_START_LINE: c_int = 108;