+++ openssl/src/pkey.rs
@@ -47,7 +47,7 @@
 
 use ffi;
 use foreign_types::{ForeignType, ForeignTypeRef};
-use libc::c_int;
+use libc::{c_int, c_long};
 use std::ffi::CString;
 use std::mem;
 use std::ptr;
@@ -90,6 +90,11 @@ impl Id {
     pub const DSA: Id = Id(ffi::EVP_PKEY_DSA);
     pub const DH: Id = Id(ffi::EVP_PKEY_DH);
     pub const EC: Id = Id(ffi::EVP_PKEY_EC);
+
+    #[cfg(ossl111)]
+    pub const ED25519: Id = Id(ffi::EVP_PKEY_ED25519);
+    #[cfg(ossl111)]
+    pub const ED448: Id = Id(ffi::EVP_PKEY_ED448);
 }
 
 /// A trait indicating that a key has parameters.
@@ -119,6 +124,17 @@ generic_foreign_type_and_impl_send_sync! {
     pub struct PKey<T>;
     /// Reference to `PKey`.
     pub struct PKeyRef<T>;
+}
+
+impl<T> ToOwned for PKeyRef<T> {
+    type Owned = PKey<T>;
+
+    fn to_owned(&self) -> PKey<T> {
+        unsafe {
+            EVP_PKEY_up_ref(self.as_ptr());
+            PKey::from_ptr(self.as_ptr())
+        }
+    }
 }
 
 impl<T> PKeyRef<T> {
@@ -177,6 +193,15 @@ impl<T> PKeyRef<T> {
     /// [`EVP_PKEY_id`]: https://www.openssl.org/docs/man1.1.0/crypto/EVP_PKEY_id.html
     pub fn id(&self) -> Id {
         unsafe { Id::from_raw(ffi::EVP_PKEY_id(self.as_ptr())) }
+    }
+
+    /// Returns the maximum size of a signature in bytes.
+    ///
+    /// This corresponds to [`EVP_PKEY_size`].
+    ///
+    /// [`EVP_PKEY_size`]: https://www.openssl.org/docs/man1.1.1/man3/EVP_PKEY_size.html
+    pub fn size(&self) -> usize {
+        unsafe { ffi::EVP_PKEY_size(self.as_ptr()) as usize }
     }
 }
 
@@ -255,6 +280,12 @@ where
         /// [`i2d_PrivateKey`]: https://www.openssl.org/docs/man1.0.2/crypto/i2d_PrivateKey.html
         private_key_to_der,
         ffi::i2d_PrivateKey
+    }
+}
+
+impl<T> Clone for PKey<T> {
+    fn clone(&self) -> PKey<T> {
+        PKeyRef::to_owned(self)
     }
 }
 
@@ -417,6 +448,40 @@ impl PKey<Private> {
         }
     }
 
+    #[cfg(ossl110)]
+    fn generate_eddsa(nid: c_int) -> Result<PKey<Private>, ErrorStack> {
+        unsafe {
+            let kctx = cvt_p(ffi::EVP_PKEY_CTX_new_id(nid, ptr::null_mut()))?;
+            let ret = cvt(ffi::EVP_PKEY_keygen_init(kctx));
+            if let Err(e) = ret {
+                ffi::EVP_PKEY_CTX_free(kctx);
+                return Err(e);
+            }
+            let mut key = ptr::null_mut();
+            let ret = cvt(ffi::EVP_PKEY_keygen(kctx, &mut key));
+
+            ffi::EVP_PKEY_CTX_free(kctx);
+
+            if let Err(e) = ret {
+                return Err(e);
+            }
+
+            Ok(PKey::from_ptr(key))
+        }
+    }
+
+    /// Generates a new private Ed25519 key
+    #[cfg(ossl111)]
+    pub fn generate_ed25519() -> Result<PKey<Private>, ErrorStack> {
+        PKey::generate_eddsa(ffi::EVP_PKEY_ED25519)
+    }
+
+    /// Generates a new private Ed448 key
+    #[cfg(ossl111)]
+    pub fn generate_ed448() -> Result<PKey<Private>, ErrorStack> {
+        PKey::generate_eddsa(ffi::EVP_PKEY_ED448)
+    }
+
     private_key_from_pem! {
         /// Deserializes a private key from a PEM-encoded key type specific format.
         ///
@@ -457,6 +522,28 @@ impl PKey<Private> {
         private_key_from_der,
         PKey<Private>,
         ffi::d2i_AutoPrivateKey
+    }
+
+    /// Deserializes a DER-formatted PKCS#8 unencrypted private key.
+    ///
+    /// This method is mainly for interoperability reasons. Encrypted keyfiles should be preferred.
+    pub fn private_key_from_pkcs8(
+        der: &[u8],
+    ) -> Result<PKey<Private>, ErrorStack>
+    {
+        unsafe {
+            ffi::init();
+            let len = der.len().min(c_long::max_value() as usize) as c_long;
+            let p8inf = cvt_p(ffi::d2i_PKCS8_PRIV_KEY_INFO(
+                ptr::null_mut(),
+                &mut der.as_ptr(),
+                len,
+            ))?;
+            let res = cvt_p(ffi::EVP_PKCS82PKEY(p8inf))
+                .map(|p| PKey::from_ptr(p));
+            ffi::PKCS8_PRIV_KEY_INFO_free(p8inf);
+            res
+        }
     }
 
     /// Deserializes a DER-formatted PKCS#8 private key, using a callback to retrieve the password
@@ -536,6 +623,22 @@ impl PKey<Public> {
     }
 }
 
+cfg_if! {
+    if #[cfg(any(ossl110, libressl270))] {
+        use ffi::EVP_PKEY_up_ref;
+    } else {
+        unsafe extern "C" fn EVP_PKEY_up_ref(pkey: *mut ffi::EVP_PKEY) {
+            ffi::CRYPTO_add_lock(
+                &mut (*pkey).references,
+                1,
+                ffi::CRYPTO_LOCK_EVP_PKEY,
+                "pkey.rs\0".as_ptr() as *const _,
+                line!() as c_int,
+            );
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use dh::Dh;
@@ -556,6 +659,12 @@ mod tests {
             .unwrap();
         PKey::private_key_from_pem_passphrase(&pem, b"foobar").unwrap();
         assert!(PKey::private_key_from_pem_passphrase(&pem, b"fizzbuzz").is_err());
+    }
+
+    #[test]
+    fn test_unencrypted_pkcs8() {
+        let key = include_bytes!("../test/pkcs8-nocrypt.der");
+        PKey::private_key_from_pkcs8(key).unwrap();
     }
 
     #[test]